pyproject.toml
[project]
name = "safer-bench"
version = "0.1.0"
description = "A comprehensive benchmark framework for evaluating Federated Retrieval-Augmented Generation (RAG) systems with approval-based privacy controls."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "syft-flwr>=0.2.1",
    "numpy>=2.3.3",
    "faiss-cpu>=1.10.0",
    "tqdm>=4.67",
    "sentence_transformers>=5.0.0",
    "jupyterlab>=4.4.7",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.1",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/safer_bench"]
only-include = ["src"]

[tool.hatch.build.targets.sdist]
only-include = ["src", "pyproject.toml", "README.md"]

[project.scripts]
safer-bench = "safer_bench.main:main"

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.3.0",
]
